refactor(ui): added prefix "ui" to ui files
