+++ src/include/ui/ui_helper_functions.h
