+++ src/ui/ui_helper_functions.c
@@ -1,4 +1,4 @@
-#include "../include/ui/helper_functions.h"
+#include "../include/ui/ui_helper_functions.h"
 #include "../../include/console.h"
 #include "../../include/helper_functions.h"
 #include <stdio.h>
